requirements.txt
Django==4.2.2
gunicorn==23.0.0
python-dotenv==1.1.1

# Database
psycopg2-binary==2.9.11

# Static files
whitenoise==6.6.0

# Media
django-cloudinary-storage==0.3.0

# Images
Pillow==10.0.0

# Utilities
requests==2.32.4
python-slugify==8.0.4

